# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: CameraEye API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and account
  - name: shop
    description: Shop inventory
  - name: orders
    description: Order creation and history
  - name: admin
    description: Admin order management
  - name: wishlist
    description: Per-account wishlist
  - name: newsletter
    description: Newsletter subscriptions
  - name: downloads
    description: Digital download delivery
  - name: instagram
    description: Instagram feed proxy
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create an account
      description: Registers a new user and starts a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created and session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      description: Verifies credentials and starts a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      description: Destroys the current session
      responses:
        "200":
          description: Session destroyed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current user
      description: Returns the authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /shop/inventory:
    get:
      operationId: getInventory
      tags: [shop]
      summary: Edition inventory
      description: Sold counts and remaining stock for limited-edition products
      responses:
        "200":
          description: Inventory status per product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      description: >
        Creates an order (guest or authenticated). Print orders start in
        status "pending" and are confirmed manually; digital-only orders are
        completed immediately and download tokens are returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid order (unknown product, bad address, ...)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Insufficient edition stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/mine:
    get:
      operationId: getMyOrders
      tags: [orders]
      summary: My order history
      description: Orders belonging to the authenticated user
      responses:
        "200":
          description: Order list, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist
      description: Product ids saved by the authenticated user
      responses:
        "200":
          description: Wishlist product ids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateWishlist
      tags: [wishlist]
      summary: Replace wishlist
      description: Replaces the authenticated user's wishlist with the given product ids
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistUpdateRequest"
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to the newsletter
      description: Adds an email to the newsletter list (idempotent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeRequest"
      responses:
        "200":
          description: Subscribed (or already subscribed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /downloads/{token}:
    get:
      operationId: redeemDownload
      tags: [downloads]
      summary: Redeem a download token
      description: >
        Exchanges a download token for the file URL. Each call counts against
        the token's download limit.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Download information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadInfoResponse"
        "404":
          description: Unknown token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Token expired or download limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /instagram/feed:
    get:
      operationId: getInstagramFeed
      tags: [instagram]
      summary: Latest Instagram posts
      description: >
        Proxies the Instagram Graph API. Returns an empty list when no
        access token is configured server-side.
      responses:
        "200":
          description: Recent posts (possibly empty)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramFeedResponse"
  /admin/orders:
    get:
      operationId: getAdminOrders
      tags: [admin]
      summary: List all orders
      description: All orders, optionally filtered by status (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: Order list, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status
      description: >
        Moves an order to a new status (admin only). Legal transitions follow
        the shared order-status flow; customers are notified by email.
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Illegal status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unknown order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        isAdmin:
          type: boolean
      required:
        - id
        - email
        - name
        - isAdmin
    AuthUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 120
      required:
        - email
        - password
        - name
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      required:
        - email
        - password
    InventoryItem:
      type: object
      properties:
        productId:
          type: string
        editionSize:
          type: integer
        sold:
          type: integer
        remaining:
          type: integer
        soldOut:
          type: boolean
      required:
        - productId
        - sold
        - soldOut
    InventoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
      required:
        - items
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - production
        - shipped
        - delivered
        - completed
        - cancelled
    ShippingAddress:
      type: object
      properties:
        recipient:
          type: string
          minLength: 1
          maxLength: 120
        country:
          type: string
          minLength: 2
          maxLength: 8
        city:
          type: string
          minLength: 1
          maxLength: 120
        postalCode:
          type: string
          minLength: 1
          maxLength: 20
        addressLine1:
          type: string
          minLength: 1
          maxLength: 200
        addressLine2:
          type: string
          maxLength: 200
        phone:
          type: string
          maxLength: 40
      required:
        - recipient
        - country
        - city
        - postalCode
        - addressLine1
    OrderItemInput:
      type: object
      properties:
        productId:
          type: string
        sizeId:
          type: string
        quantity:
          type: integer
          minimum: 1
          maximum: 10
      required:
        - productId
        - quantity
    CreateOrderRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
        items:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/OrderItemInput"
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        notes:
          type: string
          maxLength: 2000
      required:
        - name
        - email
        - items
    OrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productType:
          type: string
        title:
          type: string
        sizeId:
          type: string
        unitPrice:
          type: integer
        quantity:
          type: integer
      required:
        - id
        - productId
        - productType
        - title
        - unitPrice
        - quantity
    DownloadTokenInfo:
      type: object
      properties:
        productId:
          type: string
        title:
          type: string
        token:
          type: string
        expiresAt:
          type: string
        remainingDownloads:
          type: integer
      required:
        - productId
        - title
        - token
        - expiresAt
        - remainingDownloads
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        email:
          type: string
        name:
          type: string
        subtotal:
          type: integer
        shippingFee:
          type: integer
        total:
          type: integer
        currency:
          type: string
        trackingNumber:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        downloads:
          type: array
          items:
            $ref: "#/components/schemas/DownloadTokenInfo"
      required:
        - id
        - orderNumber
        - status
        - email
        - name
        - subtotal
        - shippingFee
        - total
        - currency
        - createdAt
        - items
    OrderResponse:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
      required:
        - order
    OrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required:
        - orders
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
        trackingNumber:
          type: string
          maxLength: 64
      required:
        - status
    WishlistResponse:
      type: object
      properties:
        productIds:
          type: array
          items:
            type: string
      required:
        - productIds
    WishlistUpdateRequest:
      type: object
      properties:
        productIds:
          type: array
          maxItems: 200
          items:
            type: string
      required:
        - productIds
    NewsletterSubscribeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    DownloadInfoResponse:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        fileSpec:
          type: string
        remainingDownloads:
          type: integer
      required:
        - title
        - url
        - remainingDownloads
    InstagramPost:
      type: object
      properties:
        id:
          type: string
        mediaUrl:
          type: string
        permalink:
          type: string
        caption:
          type: string
      required:
        - id
        - mediaUrl
        - permalink
    InstagramFeedResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/InstagramPost"
      required:
        - posts