# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CameraEye API specification
 * OpenAPI spec version: 0.2.0
 */

export interface LoginRequest {
  email: string;
  /** @minLength 1 */
  password: string;
}