# lib/api-zod/src/generated/types/newsletterSubscribeRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CameraEye API specification
 * OpenAPI spec version: 0.2.0
 */

export interface NewsletterSubscribeRequest {
  email: string;
}