# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CameraEye API specification
 * OpenAPI spec version: 0.2.0
 */
import type { DownloadTokenInfo } from './downloadTokenInfo';
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';
import type { ShippingAddress } from './shippingAddress';

export interface Order {
  id: string;
  orderNumber: string;
  status: OrderStatus;
  email: string;
  name: string;
  subtotal: number;
  shippingFee: number;
  total: number;
  currency: string;
  trackingNumber?: string;
  notes?: string;
  createdAt: string;
  shippingAddress?: ShippingAddress;
  items: OrderItem[];
  downloads?: DownloadTokenInfo[];
}