# lib/api-zod/src/generated/types/shippingAddress.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CameraEye API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ShippingAddress {
  /**
     * @minLength 1
     * @maxLength 120
     */
  recipient: string;
  /**
     * @minLength 2
     * @maxLength 8
     */
  country: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  city: string;
  /**
     * @minLength 1
     * @maxLength 20
     */
  postalCode: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  addressLine1: string;
  /** @maxLength 200 */
  addressLine2?: string;
  /** @maxLength 40 */
  phone?: string;
}