# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CameraEye API specification
 * OpenAPI spec version: 0.2.0
 */

export interface User {
  id: string;
  email: string;
  name: string;
  isAdmin: boolean;
}